move lanecover implementation to ManiaModHidden
